Don't complain about `explicit` no-arg constructors. (#227)
Previously, cpplint has complained when a no-argument constructor is
marked `explicit`. This is bad for two reasons:

1. As of C++11, converting constructors can be called with any number of
arguments including 0. Therefore, this check is wrong. (see
https://en.cppreference.com/w/cpp/language/converting_constructor )

2. Even in pre-C++-11 code, it is good defensive programming practice to
mark all constructors `explicit`. Marking a constructor that cannot be a
conversion `explicit` has no ill effect, but if someone alters it to be
a possible implicit conversion, it defends you against them forgetting
to add `explicit`.

---------

Co-authored-by: Aaron Liu <aaronliu0130@gmail.com>