Add a pyproject.toml file (#269)
By explicitly specifying the legacy setuptools build backend via
pyproject.toml, we ensure that the current directory is visible to the
setup.py invocation. Without this, in some environments I observe that
pip installing the sdist or a clone of the repo fails because the
current directory is not on the PYTHONPATH (for instance, using cpplint
as a pre-commit hook fails for me on some Python/pip versions). Note
that this is because the setup.py in this repo imports the cpplint
module from the current directory, and for this reason it is necessary
to specify that the build backend is `setuptools.build_meta:__legacy__`,
not just `setuptools.build_meta`.